Reorganize files.
