+++ client/src/components/filters/time/geologic/GeologicFormatFieldset.jsx
@@ -2,13 +2,9 @@ import React, {useState, useEffect} from 'react'
 import _ from 'lodash'
 import moment from 'moment/moment'
 
-import {FilterColors, SiteColors} from '../../../style/defaultStyles'
-import {ymdToDateMap, isValidDate} from '../../../utils/inputUtils'
+import {FilterColors, SiteColors} from '../../../../style/defaultStyles'
 
-import FilterFieldset from '../FilterFieldset'
-import YearField from './YearField'
-import MonthField from './MonthField'
-import DayField from './DayField'
+import FilterFieldset from '../../FilterFieldset'
 
 const styleDate = {
   display: 'flex',
