+++ client/src/components/filters/time/standard/DateFieldset.jsx
@@ -2,10 +2,10 @@ import React, {useState, useEffect} from 'react'
 import _ from 'lodash'
 import moment from 'moment/moment'
 
-import {FilterColors, SiteColors} from '../../../style/defaultStyles'
-import {ymdToDateMap, isValidDate} from '../../../utils/inputUtils'
+import {FilterColors, SiteColors} from '../../../../style/defaultStyles'
+import {ymdToDateMap, isValidDate} from '../../../../utils/inputUtils'
 
-import FilterFieldset from '../FilterFieldset'
+import FilterFieldset from '../../FilterFieldset'
 import YearField from './YearField'
 import MonthField from './MonthField'
 import DayField from './DayField'
